backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
python-multipart>=0.0.9
pydantic>=2.9.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
joblib>=1.4.0
requests>=2.32.0
python-magic-bin>=0.4.14

# Database dependencies (Azure PostgreSQL)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
